Create 11_All_inclusive.ipynb
What we aim for with these final exercises is to put into practice everything we have learned, except for lists, which were the focus of most exercises. However, these exercises help us significantly to reinforce and consolidate the knowledge acquired, achieving a solid understanding of what is required in basic Python.